Do not require ugly _2, _3 suffixes for node class names.
This was making a lot of code unnecessarily unclear. Especially since this was mainly to distinguish from VRML 1.0, and VRML 1.0 is now really ancient and not interesting to most people.

Also, fix SaveVRML call in sunny_street_process.lpr.

Also, make ForceConvertingToX3D default false, to fix compilation of various programs and examples.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle@9544 afe8dcc5-9094-451b-8773-5f495042a737