Add DoProgress parameter to PrepareCore, make progress optional.
This is important if Prepare happens in the middle of another progress, e.g. when you forgot to add initial creature to <resources> in level.

Remove AnimationName param and related log messages, not terribly useful and were complicating code.

GameLevel merely warns when not CreatureKind.Prepared. It will be handled 100% Ok now.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle@11176 afe8dcc5-9094-451b-8773-5f495042a737