Load sounds from XML as set as you set SoundsXmlFileName.
This allows you to set SoundsXmlFileName and use SoundFromName immediately after. No need to wait for OpenAL context initialization just to do SoundFromName. And this in turn makes sounds initialization in castle and rift trivial. No need for custom TXmlSoundEngine descendant.

This was perfectly trivial to achieve, as our ReadSounds was already designed to load sounds without AL context...


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle@11509 afe8dcc5-9094-451b-8773-5f495042a737